GGM-Dicord-bot-main/requirements.txt
discord.py==1.5.1
time.py
openpyxl.py
sys.py
